test: move generated code tests to internal
