+++ internal/defaults_test.go
@@ -1,4 +1,4 @@
-package ogen
+package internal
 
 import (
 	"fmt"
