+++ internal/requests_test.go
@@ -1,4 +1,4 @@
-package ogen_test
+package internal_test
 
 import (
 	"context"
